+++ test/typescript/apis/project/cds-ql.ts
@@ -1,8 +1,20 @@
-import { ArrayConstructable } from '../../../../apis/internal/inference'
-import { QLExtensions_ } from '../../../../apis/internal/query'
-import { DeepRequired } from '../../../../apis/internal/util'
 import { QLExtensions } from '../../../../apis/ql'
 import { Foo, Foos, attach } from './dummy'
+
+// @ts-expect-error - only supposed to be used statically, constructors private
+new SELECT;
+// @ts-expect-error
+new INSERT;
+// @ts-expect-error
+new UPDATE;
+// @ts-expect-error
+new UPSERT;
+// @ts-expect-error
+new DELETE;
+// @ts-expect-error
+new CREATE; 
+// @ts-expect-error
+new DROP;
 
 // unwrapped plural types
 let sel: SELECT<Foos>